Basic Spring Security involved with spring security starter
Here you can see when we add dependency for spring security and hit application url we face a login page for which username is user and password is generated in console. Once login JSession cookie will be generated with which we don't need to login again for hitting other urls. Once cookie is cleared we again need to login.